Initial Protocol - client
Commit Summary
==============
I did this in coffee and converted it to js on the fly. I also used
packackes in the node_modules folder for testing. In particular, I used
mocha-phantomjs and chai. I felt a headless browser would be fitting,
seeing how the VegaClient should be put on the `window` global.